Move markdown writing code
Move core markdown writing code into its own library, attempting to strip lib.rs into only the components we'll expose to the command-line tool.
